refactor: usecase
