+++ domain/usecases/account/usecase.go
@@ -6,10 +6,20 @@ import (
 	"github.com/Vivi3008/apiTestGolang/domain/entities/account"
 )
 
+type AccountUsecase struct {
+	repo account.AccountRepository
+}
+
 type Usecase interface {
 	CreateAccount(ctx context.Context, account account.Account) (account.Account, error)
 	ListAllAccounts(ctx context.Context) ([]account.Account, error)
 	ListAccountById(ctx context.Context, accountId string) (account.Account, error)
 	NewLogin(ctx context.Context, login account.Login) (string, error)
 	UpdateAccountBalance(ctx context.Context, accountId string, value int, method MethodPayment) (account.Account, error)
+}
+
+func NewAccountUsecase(acc account.AccountRepository) AccountUsecase {
+	return AccountUsecase{
+		repo: acc,
+	}
 }
