+++ domain/usecases/activities/usecase.go
@@ -7,3 +7,10 @@ import (
 type Usecase interface {
 	ListActivity(ctx context.Context, accountId string) ([]AccountActivity, error)
 }
+type ActivityUsecase struct {
+	actRepo AccountActivityRepository
+}
+
+func NewAccountActivityUsecase(act AccountActivityRepository) ActivityUsecase {
+	return ActivityUsecase{actRepo: act}
+}
